requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
certifi==2025.10.5
click==8.3.0
granian==2.5.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
redis==6.4.0
reflex==0.8.15
reflex-hosting-cli==0.1.59
rich==14.2.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.2.0
